moved raw serializer to test
