+++ common/src/Microsoft.Azure.IIoT.Serializers.NewtonSoft/src/Default/NewtonSoftJsonSerializer.cs
@@ -535,29 +535,6 @@ namespace Microsoft.Azure.IIoT.Serializers.NewtonSoft {
         }
 
         /// <summary>
-        /// Json raw converter
-        /// </summary>
-        public sealed class RawJsonConverter : JsonConverter {
-
-            /// <inheritdoc/>
-            public override bool CanConvert(Type objectType) {
-                return true;
-            }
-
-            /// <inheritdoc/>
-            public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer) {
-                // convert parsed JSON back to string
-                return JToken.Load(reader).ToString(Formatting.None);
-            }
-
-            /// <inheritdoc/>
-            public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer) {
-                // write value directly to output; assumes string is already JSON
-                writer.WriteRawValue((string)value);
-            }
-        }
-
-        /// <summary>
         /// Strategy to only camel case dictionary keys
         /// </summary>
         private class CamelCaseDictionaryKeys : CamelCaseNamingStrategy {
